Moving timing table on the OperationTimers class
