+++ src/synthesizer/pipeline/pipeline_utils.py
@@ -21,7 +21,9 @@ from synthesizer.instruments import InstrumentCollection
 from synthesizer.photometry import PhotometryCollection
 from synthesizer.synth_warnings import warn
 from synthesizer.units import unit_is_compatible
-from synthesizer.utils.operation_timers import timed
+from synthesizer.utils.operation_timers import (
+    timed,
+)
 
 # Special model label for operations that are not tied to a specific model.
 # This can be used for operations such as SFZH / SFH with no relation to
@@ -276,194 +278,6 @@ def combine_atomic_timing_snapshots(
     # Sum the per-rank elapsed times so the overhead contribution is measured
     # in the same rank-seconds units as the atomic timing totals.
     return combined, sum(gathered_elapsed)
-
-
-def build_timing_analysis_rows(timing_data, total_elapsed):
-    """Build sorted timing rows including overhead and total entries.
-
-    Args:
-        timing_data (dict):
-            A timing dictionary keyed by operation name with ``seconds``,
-            ``count``, and ``source`` entries for each operation.
-        total_elapsed (float):
-            The total elapsed wall-clock time represented by the analysis.
-
-    Returns:
-        list:
-            A list of row dictionaries sorted by descending operation time,
-            with trailing ``Overhead`` and ``Total`` summary rows appended.
-    """
-    # Account for any elapsed time that is not represented in the atomic timer
-    # totals so the report shows both timed and overhead fractions explicitly.
-    timed_elapsed = sum(data["seconds"] for data in timing_data.values())
-    untimed_elapsed = max(total_elapsed - timed_elapsed, 0.0)
-
-    # Build one row per timed operation, including its contribution as a
-    # percentage of the total elapsed time.
-    rows = []
-    for operation, data in timing_data.items():
-        fraction = (
-            data["seconds"] / total_elapsed * 100.0
-            if total_elapsed > 0.0
-            else 0.0
-        )
-        rows.append(
-            {
-                "operation": operation,
-                "seconds": data["seconds"],
-                "fraction_percent": fraction,
-                "count": data["count"],
-                "source": data["source"],
-            }
-        )
-
-    # Present the operations in descending order of time spent so the most
-    # important contributors are shown first in both tables and plots.
-    rows.sort(key=lambda row: row["seconds"], reverse=True)
-
-    # Append explicit summary rows so the overhead contribution and total are
-    # available to both the terminal table and the written CSV file.
-    rows.append(
-        {
-            "operation": "Overhead",
-            "seconds": untimed_elapsed,
-            "fraction_percent": (
-                untimed_elapsed / total_elapsed * 100.0
-                if total_elapsed > 0.0
-                else 0.0
-            ),
-            "count": None,
-            "source": "N/A",
-        }
-    )
-    rows.append(
-        {
-            "operation": "Total",
-            "seconds": total_elapsed,
-            "fraction_percent": 100.0 if total_elapsed > 0.0 else 0.0,
-            "count": None,
-            "source": "N/A",
-        }
-    )
-
-    return rows
-
-
-def print_timing_analysis_table(rows, print_func=print):
-    """Print a timing analysis table to stdout.
-
-    Args:
-        rows (list):
-            The timing rows produced by ``build_timing_analysis_rows``.
-        print_func (callable):
-            The function used to emit each formatted line. This defaults to the
-            built-in ``print`` but can be replaced with ``Pipeline._print`` to
-            match the standard pipeline logging style.
-
-    Returns:
-        None
-    """
-    # Filter the printed table to operations that contribute at least 0.01% of
-    # the total elapsed time so the displayed precision matches the fraction
-    # column. Always keep the synthetic summary rows.
-    filtered_rows = []
-    for row in rows:
-        if row["operation"] in ("Overhead", "Total"):
-            filtered_rows.append(row)
-            continue
-
-        if row["fraction_percent"] >= 0.01:
-            filtered_rows.append(row)
-
-    # Pre-format the row values so the final column widths reflect the exact
-    # strings that will be printed rather than just the raw underlying values.
-    formatted_rows = []
-    for row in filtered_rows:
-        # Show sub-centisecond timings in scientific notation so tiny but still
-        # visible entries are distinguishable in the printed table.
-        if 0.0 < row["seconds"] < 0.01:
-            seconds_str = f"{row['seconds']:.2e}"
-        else:
-            seconds_str = f"{row['seconds']:.2f}"
-
-        formatted_rows.append(
-            {
-                "operation": row["operation"],
-                "seconds": seconds_str,
-                "fraction_percent": f"{row['fraction_percent']:.2f}",
-                "count": "-" if row["count"] is None else str(row["count"]),
-                "source": row["source"],
-            }
-        )
-
-    # Compute column widths from the fully formatted data so every column is
-    # wide enough for both its header and its largest row value.
-    operation_width = max(
-        len("Operation"), *(len(r["operation"]) for r in formatted_rows)
-    )
-    time_width = max(
-        len("Time (s)"), *(len(r["seconds"]) for r in formatted_rows)
-    )
-    frac_width = max(
-        len("Fraction (%)"),
-        *(len(r["fraction_percent"]) for r in formatted_rows),
-    )
-    count_width = max(len("Count"), *(len(r["count"]) for r in formatted_rows))
-    source_width = max(
-        len("Source"), *(len(r["source"]) for r in formatted_rows)
-    )
-
-    # Construct a single divider string that can be reused for the header,
-    # body separator, and closing line.
-    divider = (
-        "+"
-        + "-" * (operation_width + 2)
-        + "+"
-        + "-" * (time_width + 2)
-        + "+"
-        + "-" * (frac_width + 2)
-        + "+"
-        + "-" * (count_width + 2)
-        + "+"
-        + "-" * (source_width + 2)
-        + "+"
-    )
-
-    # Print the table header with the derived column widths.
-    print_func(divider)
-    print_func(
-        "| "
-        + f"{'Operation':<{operation_width}}"
-        + " | "
-        + f"{'Time (s)':>{time_width}}"
-        + " | "
-        + f"{'Fraction (%)':>{frac_width}}"
-        + " | "
-        + f"{'Count':>{count_width}}"
-        + " | "
-        + f"{'Source':<{source_width}}"
-        + " |"
-    )
-    print_func(divider)
-
-    # Print each timing row, converting missing counts on summary rows into a
-    # simple placeholder for readability.
-    for row in formatted_rows:
-        print_func(
-            "| "
-            + f"{row['operation']:<{operation_width}}"
-            + " | "
-            + f"{row['seconds']:>{time_width}}"
-            + " | "
-            + f"{row['fraction_percent']:>{frac_width}}"
-            + " | "
-            + f"{row['count']:>{count_width}}"
-            + " | "
-            + f"{row['source']:<{source_width}}"
-            + " |"
-        )
-
-    print_func(divider)
 
 
 def write_timing_analysis_summary(rows, outdir):
